Component::get_ functions can use ExportLookup as well now
